requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.38.1
bleach==6.2.0
cachetools==5.5.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
cityscapesScripts==2.2.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.6
distro==1.9.0
diversipy==0.9
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.1.1
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.4
inflect==6.0.2
ipykernel==6.22.0
ipython==8.11.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.2.0
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.13
keras==2.11.0
kiwisolver==1.4.7
kornia==0.6.9
libclang==18.1.1
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.6.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
num2words==0.5.12
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.1
opencv-python-headless==4.5.5.64
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.2
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.6
prettytable==3.6.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.19.6
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.1.1
pycocotools==2.0.6
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0
pyzmq==26.2.0
qd==0.8.9
qtconsole==5.6.1
QtPy==2.4.2
referencing==0.35.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scipy==1.9.3
Send2Trash==1.8.3
sentry-sdk==2.18.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
timm==0.6.12
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.1.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@13493215abceafc1653af88b045120014fb4c1fc
types-python-dateutil==2.9.0.20241003
typing==3.7.4.3
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.32.1
uvloop==0.21.0
wandb==0.13.9
watchfiles==1.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
word2number==1.1
wrapt==1.16.0
yacs==0.1.8
yarl==1.18.0
